feature(sender): adapted relay and sender component to work with a simple transfer name
In order to be able to use a simple transfer name, the relay server and
the sender were adapted. On the sender, the on_create_room method was
adapted so that a call is made to the http_client to register the
transfer with the relay server. The handle_create_room method on the
server has been adapted so that when a room is created, the name is
transmitted by the receiver and is no longer generated on the
relayserver.This was also done in preparation for integrating a local
relayserver for the direct connection between sender and receiver.